requirements.txt
appdirs==1.4.3
certifi==2020.4.5.1
distlib==0.3.0
filelock==3.0.12
gunicorn==20.0.4
pipenv==2018.11.26
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-restframework==0.0.1
djangorestframework==3.11.0
pip==20.0.2
pytz==2019.3
setuptools==46.1.3
sqlparse==0.3.1
wheel==0.34.2
whitenoise==5.0.1
